backend/map_trees/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.14
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
python-dotenv==0.20.0
geopy==2.2.0
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
Pillow==9.1.1